requirements.txt
#pip install -r requirements.txt
#create : virtualenv anc
#activate : anc\Scripts\activate
#deactive : deactivate
pillow
matplotlib
numpy
tensorflow
tensorflow_hub
pywebio
